Perfectly working condition
This is a program where I practised OOPS in python. Here, I defined a class which make up the structure of the railway ticket form and then the passengers which are the objects in our case. This takes as many passengers you want in your ticket (I didn't cap it to 6) and then take more necessary details and then book ticket for you. The ticket is then saved in your folder where your travel details can be found...WHY DON' T YOU GIVE A TRY?